[Fix] #86 - 구글맵API 관련 파일들 주석처리해서 Resource-GoogleMapService로 옮기기
